Split image and utils, use utils from site-react, add image implementation for plain react
